Z index works

+++ PixForge/Platforms/Interfaces/iRenderer.h
@@ -16,7 +16,7 @@ public:
 public:
   virtual void render() = 0;
   virtual void renderBackground(STL::Vec<float, 4> colour) = 0;
-  virtual void renderColourBox(STL::Vec<int, 5> position, STL::Vec<float, 4> colour) = 0;
+  virtual void renderColourBox(STL::Vec<float, 5> position, STL::Vec<float, 4> colour) = 0;
   virtual void renderTexture() = 0;
   virtual void bindVertexShader(PF::ENGINE::Shader shader) = 0;
   virtual void bindFragmentShader(PF::ENGINE::Shader shader) = 0;
